ocr/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
flask-cors==4.0.0
python-multipart==0.0.6
Pillow==10.0.0
gevent==23.9.1